+++ src/main/kotlin/com/cashtrack/cashtrack_api/domain/dto/response/UserResponse.kt
@@ -1,7 +1,8 @@
-package com.cashtrack.cashtrack_api.domain.dto
+package com.cashtrack.cashtrack_api.domain.dto.response
 
 import java.io.Serializable
 
+@Suppress("SerialVersionUIDInSerializableClass")
 data class UserResponse(
     val id:Long?,
     val username:String
