+++ src/main/kotlin/com/cashtrack/cashtrack_api/port/controller/UserController.kt
@@ -1,9 +1,9 @@
 package com.cashtrack.cashtrack_api.port.controller
 
 import com.cashtrack.cashtrack_api.application.service.UserService
-import com.cashtrack.cashtrack_api.domain.auth.request.UserRegisterRequest
-import com.cashtrack.cashtrack_api.domain.auth.response.BalanceResponse
-import com.cashtrack.cashtrack_api.domain.auth.response.UserResponse
+import com.cashtrack.cashtrack_api.domain.dto.request.UserRegisterRequest
+import com.cashtrack.cashtrack_api.domain.dto.response.BalanceResponse
+import com.cashtrack.cashtrack_api.domain.dto.response.UserResponse
 import io.swagger.v3.oas.annotations.security.SecurityRequirement
 import jakarta.transaction.Transactional
 import jakarta.validation.Valid
